FEATURE: Implement verify method in regex analyzer
That verify method can be used to perform further checks after a subclass of the regex analyzer receives the results from the analyzer. In the subclass we can now for example check if we received a minimum amount of results or if the results contains certain values or has a certain length.