fix(actions): store each paste in own file
Previously the SaveFileAction did store (and overwrite) all pastes in a single file - or it would crash, because the path parameter was a directory.
This commit fixes that behaviour and also adds the usage of the TemplatingEngine to have better control about what should be saved.

Apart from that, it's now also possible to choose a custom file_ending.

fixes #179