feat(ircaction): implement high quality version of ircaction
The previous implementation lacked some features such as session keepalive, various checks (message length), etc.
In this implementation we setup the IRC connection once at initialization of the action and then run the irc client in a different thread. It automatically reconnects after the connection has been lost, replies to server PINGs and much more.

I did not add a lot of error handling, so it might be possible that the IRC action shows unexpected behaviour when starting without a proper connection. There could also be more logging for when a connection was not successful, so that users can more easily identify issues with their connection or configuration.

Also passwords and ssl are both not implemented yet.