Documention how to download pastepwn
While downloading the software, I made a couple of screenshots of some specific code that needs to be used to activate the software. I also noticed how complex it can be if you need to learn how to download the software. The main work I want to complete is a full guide on how to download/troubleshoot all the problems so all users can install and run the software efficiently.